+++ api/api.go
@@ -23,12 +23,6 @@ type HTTPHandler struct {
 	Method  string // The assoicated HTTP method
 }
 
-// Sum contains digests for a certificate or certificate request.
-type Sum struct {
-	MD5  string `json:"md5"`
-	SHA1 string `json:"sha-1"`
-}
-
 // HandlerFunc is similar to the http.HandlerFunc type; it serves as
 // an adapter allowing the use of ordinary functions as Handlers. If
 // f is a function with the appropriate signature, HandlerFunc(f) is a
@@ -212,7 +206,7 @@ func NewErrorResponse(message string, code int) Response {
 
 // sendResponse builds a response from the result, sets the JSON
 // header, and writes to the http.ResponseWriter.
-func sendResponse(w http.ResponseWriter, result interface{}) error {
+func SendResponse(w http.ResponseWriter, result interface{}) error {
 	response := NewSuccessResponse(result)
 	w.Header().Set("Content-Type", "application/json")
 	enc := json.NewEncoder(w)
