+++ api/generator/generator_test.go
@@ -1,0 +1,70 @@
+package generator
+
+import (
+	"bytes"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/cloudflare/cfssl/csr"
+)
+
+func csrData(t *testing.T) *bytes.Reader {
+	req := &csr.CertificateRequest{
+		Names: []csr.Name{
+			{
+				C:  "US",
+				ST: "California",
+				L:  "San Francisco",
+				O:  "CloudFlare",
+				OU: "Systems Engineering",
+			},
+		},
+		CN:    "cloudflare.com",
+		Hosts: []string{"cloudflare.com"},
+		KeyRequest: &csr.KeyRequest{
+			Algo: "ecdsa",
+			Size: 256,
+		},
+	}
+	csrBytes, err := json.Marshal(req)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return bytes.NewReader(csrBytes)
+}
+
+func TestGeneratorRESTfulVerbs(t *testing.T) {
+	handler, _ := NewGeneratorHandler(CSRValidate)
+	ts := httptest.NewServer(handler)
+	data := csrData(t)
+	// POST should work.
+	req, _ := http.NewRequest("POST", ts.URL, data)
+	resp, _ := http.DefaultClient.Do(req)
+	if resp.StatusCode != http.StatusOK {
+		t.Fatal(resp.Status)
+	}
+
+	// Test GET, PUT, DELETE and whatever, expect 400 errors.
+	req, _ = http.NewRequest("GET", ts.URL, data)
+	resp, _ = http.DefaultClient.Do(req)
+	if resp.StatusCode != http.StatusMethodNotAllowed {
+		t.Fatal(resp.Status)
+	}
+	req, _ = http.NewRequest("PUT", ts.URL, data)
+	resp, _ = http.DefaultClient.Do(req)
+	if resp.StatusCode != http.StatusMethodNotAllowed {
+		t.Fatal(resp.Status)
+	}
+	req, _ = http.NewRequest("DELETE", ts.URL, data)
+	resp, _ = http.DefaultClient.Do(req)
+	if resp.StatusCode != http.StatusMethodNotAllowed {
+		t.Fatal(resp.Status)
+	}
+	req, _ = http.NewRequest("WHATEVER", ts.URL, data)
+	resp, _ = http.DefaultClient.Do(req)
+	if resp.StatusCode != http.StatusMethodNotAllowed {
+		t.Fatal(resp.Status)
+	}
+}
